security: successfully resolve pending pool double-spend vulnerabilities and enforce integer types

+++ ledger/ledger.go
@@ -2,13 +2,13 @@ package ledger
 
 // Transaction represents a simple value transfer.
 type Transaction struct {
-	Sender    string  `json:"sender"`
-	Recipient string  `json:"recipient"`
-	Amount    float64 `json:"amount"`
+	Sender    string `json:"sender"`
+	Recipient string `json:"recipient"`
+	Amount    int64  `json:"amount"`
 }
 
 // NewTransaction creates a new transaction instance.
-func NewTransaction(sender, recipient string, amount float64) Transaction {
+func NewTransaction(sender, recipient string, amount int64) Transaction {
 	return Transaction{
 		Sender:    sender,
 		Recipient: recipient,
